sync lume, kits in ver.20250312

+++ lume/LumeMeta/src/call_context.cpp
@@ -52,7 +52,7 @@ DefaultCallContext& DefaultCallContext::operator=(const DefaultCallContext& othe
     }
     params_.clear();
     params_.resize(other.params_.size());
-    for (int i = 0; i != params_.size(); ++i) {
+    for (size_t i = 0; i != params_.size(); ++i) {
         params_[i].name = other.params_[i].name;
         if (auto p = interface_cast<ICloneable>(other.params_[i].value)) {
             params_[i].value = interface_pointer_cast<IAny>(p->GetClone());
